feat: use a dedicated logger namespace rather than the root logger  (#152)
This makes it easier to (re)use CWAC programmatically since other
loggings won't get included in the CWAC output and parent loggers can be
configured to do different things with the CWAC namespace without
impacting the writing of the logs to file e.g. having them not printed
to stdout.

Note I don't think this unblocks running multiple instances of the CWAC
class at the same time, since the namespace is global (though it might
actually be possible with a fancy parent log handler, but I've not dug
that deeply into it)