fix: don't consider urls filtered out due to content type as having an unexpected response code (#175)
This moves our content type check to happen as part of checking the
status code rather than as part of resolving the final url & co so that
such filtering no longer gets incorrectly reported as having an
unexpected response code.

Eventually we should probably consolidate these areas of code as I think
having the logic split like it is now creates a weird contention as
evident by the names (i.e. we're checking the headers of the
headers..?), which this is a good first step towards

Resolves #165
Resolves #122